% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysql-package.R
\docType{package}
\name{lazysql}
\alias{lazysql}
\alias{lazysql-package}
\title{Lazy SQL programming}
\description{
Helper functions to build SQL statements
 for dbGetQuery or dbSendQuery under program control.
}
\details{
More patterns will be added as required.
}
\author{
Uwe Block
}
\seealso{
\code{\link{date_between}},
 \code{\link{in_condition}},
 \code{\link{natural_key}}
}

