% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_identifier_regex}
\alias{valid_identifier_regex}
\title{Regex pattern for valid SQL identifier names}
\usage{
valid_identifier_regex()
}
\value{
Character string with regular expression
}
\description{
Returns a regular expression for valid unquoted SQL identifiers.
The identifier must start with an alphabetic character followed by
alphanumeric characters or the underscore \code{_}.
}
\examples{
lazysql::valid_identifier_regex()
}
\author{
Uwe Block
}
\references{
ORACLE Database SQL Language Reference
}

